Partial Infrastructure
This version contains more of the parts needed for building a
datacenter, but some of them are still subs.  The big problem is that
once I’ve created a datacenter, I can’t do anything with it.  The
problem is with lifetimes.  Creating the datacenter requires a mutable
borrow, but the reference is still in scope when I try to print or
clone the object.  I think it has something to do with Link having a
reference to Port, which is part of NalCell, which is part of
Datacenter.  Hence, I can reach port from Link->&Port and
Datacenter->NalCell->Port, which I think is why I’m having the lifetime
problem.