Another partial version
This version removes the need for RefCell in the NalCell array.  The
trick is to split the array and pick one cell from each part for
getting a free port when connecting a link.  This way the compiler
knows that the two cells are different and allows them both to be
mutable.