Names as &'static str
In earlier versions, names were a fixed length array of chars, which
meant they had a maximum size.  This version doesn’t have that
restriction, but it does have a slow memory leak as TenantIDs are
dropped.  I’ll be replacing this version with one that uses String, but
I’m committing this version, so I don’t lose it.