Moved Routing Table to Packet Engine
The routing table is no longer shared between the Cell Agent and the
Packet Agent.  Following this commit, I’m going to have the Cell Agent
construct routing table entries from its traph data and send the
entries to the Packet Engine, which will put them in the routing table.