Compiles with threads
But it won’t compile if I actually try to start the threads.  I’m
committing this version as a baseline I can go back to as I experiment
with ways to make it actually work.

One big problem is lifetimes.  Threads can live past anything that
spawns them, so the compiler says that anything captured with move must
have a ‘static lifetime.  Instead of doing that, I’m using crossbeam,
which lets me set an explicit lifetime.   On the other hand, the things
using threads - NalCell, CellAgent, PacketEngine, Port - all live for
the lifetime of the datacenter, so ‘static should be OK.  I just can’t
figure out how to specify it.