Add entries to routing table
I can now send routing table entries from the cell agent to the packet
engine and store them in the routing table.  The trick was to clone
self.routing_table.  You’d think that updating the clone wouldn’t
update the real routing table, but it does.  That’s because the entries
are Arc<>, which means I’m cloning the references to the entries, not
the entries themselves.  Hence, the update is applied to the underlying
value, which is the same in the table and its clone.