Port talks to Cell Agent
When a port gets connected or disconnected from a link, it now sends a
status message to the Cell Agent.  The Cell Agent updates the mask for
the ConnectedPorts tree and sends the update to the Packet Engine,
which updates the corresponding entry in the routing table.

The masks aren’t right, but I’m committing this version before I debug
the problem.