Doesn't work because link channel not ready
I can get packets from the cell agent to the port, and I can have the
port wait for those packets.  However, in this version, the port starts
listening before the channel to the link is established.

I have to change when the port starts listening.  I’m freezing this
version so I have something sort of stable to come back to when I mess
up.