Changed Packet Header
Bruce pointed out that the I need the total message size when
assembling a message from packets, but I wasn’t providing that.

I now have a size field and a flag in the header.  The flag says
whether or not this is the last packet.  If so, the size is the number
of message bytes in the packet.  If not, the size is the number of
packets in the message.

I’m using a u16 for the size field, which limits the maximum message
size to 16 K times the size of our largest packets, which is roughly 8
K.  Hence, we can send messages up to 1/2 GB.

Note that I’m not doing anything special for streaming.  For now, I’m
leaving that to the applications, but we might be able to improve
performance if we know the packets are part of a stream.  I could add a
flag to the header, but I don’t know what to do with it, so I’m leaving
it out for now.