Changed unused Result from warning to compiler error
I think Rust chose the wrong default.  It lets you propagate Error
through a call chain if you wrap each call that can return a result in
a try!.  However, if you forget to, the compiler gives a warning.  The
result is a crashed thread with no error reporting.  You can make the
warning an error, which is what I’ve done in this version.

This version also includes some fixes where I was picking up the wrong
table index, which resulted in lost messages.