Before I try the next fix
This version gets most, but not all of the Discover messages to their
destination.  I’ve figured out one problem, but I don’t know if it’s
the only one.  I’m freezing this version, so I have something to come
back to when I screw up the fix.