Found the error
Not all messages are being received, because the port closes the status
channel after reporting it is connected.  That causes the port  status
loop in the cell agent to exit, but that’s where I’m receiving messages.

A quick fix would put the packet processing in a different loop than
the handling port status, but the port will still need to report when
it gets disconnected.  I’m going to work on the port sending status
from a loop.