Minor correction to traph state
The previous version set the hop count to 0 and the path to None for a
child, which isn’t a problem for this version since those values are
meaningless for a child.  However, they are needed following a
failover.  This version sets hops to the parent hops + 1 and the path
to the parent path.  We might want to change the path later if we
decide to use more than just the root port number for the path.