Simple GVM framework in place
I can now specify a simple GVM equation, such as “true” or “false” to
turn on port 0 for the tree.  Code is in place to include local
context, such as number of hops, but it is untested.

Other code cleanup:  Changed children from Vec to HashMap to make sure
each port appears only once.

Changed compound message channels from Option to Enum to be Rustier.