UpTrees as special case
This version has a border cell create an up tree by starting a VM which
starts one or more containers.  The problem with this approach is that
we want a bunch of containers within a single VM to communicate over
trees.  This version of the code requires them to go to their
respective VM to reach other containers.  Similarly, we might want a
set of VMs to communicate with each other over trees, but this version
of the code requires them to communicate via the cell agent.

I’m freezing this version, because I’m about to try a completely
different approach in an attempt to address the issues.