Preparing for uptrees as recursive datacenters
Each uptree will be treated as a datacenter, complete with border cells
and border ports.  In preparation, I changed how I initialize the
Datacenter and Noc.

I also fixed the way I kept main from exiting and killing all my
threads.  I now do a join on each of the thread handles for the links.
The theory is that everything is dead when the last link is broken.
That makes sense, since there’s then no way to get any commands to the
cells to do anything even if they’re up.