Changes how GVM equation deserializes values
The GVM equation takes variables of a variety of types, which I pass in
via a HashMap.  I should be able to use Any in the declaration of that
map, but I can’t make it compile.  What I did instead is convert the
value to a string and put the string in the hash map.  I then
deserialize to get the value back.  (See, I told you Any would be
better.)

The previous version compiles, but it’s using a deprecated feature that
lets serde infer the type.  Given the way Rust is updated, that feature
might vanish at any moment.

This version declares an enum with the types that GvmEquation
understands.  I use those types when I deserialize to get the value.
It’s a real kludge for simple things, like hops, but I bet Any won’t
work for actual objects, while this approach will.