Added UUID to packet and table entry
Paul is worried about bit errors on the wire changing the table index
in the packet.  I added a UUID to the tree id, which I put in both the
packet header and the forwarding table entry.  The packet engine
verifies that they match for the specified table entry and throws an
error if they don’t.