Save messages for late port connections
I was doing this, but only for Discover messages.  I now do it for
messages of any kind.  In the process of making the change, I save the
packets of the message instead of the message.  That means I don’t have
to re-serialize and re-packetize every time I forward a message on the
new port.