Added a third equation to GVM
I had separate GVM equations for the cell’s ability to send and receive
on the specified tree.  I added one to denote that the message
containing the GVM equation should be saved for late port connections.
I did that implicitly for Discover messages, but there are others that
may need to be saved.