Fixed "race condition" in stacking trees
It wasn’t a race condition at all.  In order to avoid a lifetime issue,
I remove an entry from the traphs hash map, work on it, and put it
back.  When stacking a tree, I removed the black tree entry, worked on
it, and put it back with the key of the stacked tree, not the black
tree.