Changed handling of hash map entries
hash_map.get(key) gives a reference to the element, which leads to lots
of lifetime issues.  My workaround was to remove the element, work on
it, then put it back.

It turns out that Rust has Entry, which is specifically for this
purpose.  That’s what I’m using now, except for packet assemblers,
where I really do want to remove the item.

Also, minor changes to message.rs and gvm_equation.rs.