Changed GVM equation creation
I was taking four strings for the four GVM equations, but it was too
easy to get them in the wrong order.  I’m now using an enum to define
the set of equations and a hashset to pass them in.  If I forget to
specify one, it defaults to “false.”