Create output for display purposes
This version writes a file with sufficient data to animate the discover
and stack tree processes.  The file contains 3 kinds of entries.

Tree provides a mapping between a short string unique among the cells
and the cell’s UUID.

LinkID tells which ports of each cell are connected to which ports of
other cells.

Entry gives the routing table entries for each cell, which shows the
parent/child relationships.  A special entry for the Connected
pseudo-tree shows which ports are connected for each cell.