Created a Blueprint plus some cleanup
If each up tree is represented as a datacenter, I need a way to specify
the cells and links in a more general way than just a loop in main.rs.
I now have a Blueprint struct that lets you specify how many ports for
each cell, which cells are interior vs. border, and which ports on the
border cells are border ports.

I also did some cleanup to reduce the number of warnings from 600> to <
150.  The change isn’t important, but I needed to keep my fingers busy
while my brain worked on figuring out what the Blueprint should look
like.