Specify Up Tree Components
I defined a struct, DeploymentSpec, that defines how a set of VMs in a
deployment are connected, and how the containers in those VMs are
connected.

I included error checking for every specification error I can think of,
but I hope these objects will be created from a GUI.

The two .json files are my fooling around with the formats.  I’m
including them as an explanation of what the code should produce.