Changed meaning of cell type and output format
This update contains two, independent set of changes.  The simpler one
is the output for the visualization tool.  The output is now easier to
parse because it is one JSON object per line instead of a mixture of
JSON and pure text.

The second change is the meaning of CellType.  I had thought that an up
tree would consist of a datacenter consisting of VMs, and an up tree
from a VM would be a datacenter consisting of containers.  While
elegant to describe, the approach was awkward to code.  (That may just
be because I didn’t plan for it from the start.)

Instead, I’ve repurposed CellType to tell whether the cell is a border
cell or an interior cell, which saves me a mysterious true/false
parameter in some calls.