Added cell configuration enum
It is useful if you can specify the requirements on the physical cell
when deploying an up tree.  I created an enum consisting of Small,
Medium, and Large to represent this configuration.  An up tree manifest
specifies one of these options, and the GVM makes sure it gets deployed
on cells that satisfy the requirements.