Instantiated VM on up tree
This version uses the manifest to deploy a singe VM on an up tree with
a channel to the cell agent.  There’s still a lot to do, such as
setting up the map of trees the VM can talk to and wiring up a bunch of
VMs into a tree.

I also removed the need for the tree id in the packet header by
returning the table index to the cell agent.  My long term goal is to
either eliminate the tree id in the packet header to reduce latency (1
ns/bit on a 1 Gb/s link) or at least reduce it to a modest size hash,
say 16 bits, that we can use to check for bit flips on the wire.