Some changes to starting VM on an uptree
I started this version with the idea that the cell agent would have a
thread for each horizontal tree each VM could use.  However, the
consensus is that, while there won’t ever be enough allowed trees to
cause a problem, limiting the number is an unnecessary constraint this
early in the design process.  Hence, I will change so that there is one
thread to the cell agent for all allowed trees.

I’m saving this version for posterity and in case there’s something
else I want to go back to.