Fixed a problem with message dispatch
My original message dispatch only worked when the cell agent was the
recipient.  Ouch!  Fixing that problem requires setting up a proper
message dispatch.  Unfortunately, doing that right would have been a
BIG change to the code, so I came up with a kludge embodied in this
version.  Take a look at the Message.process and Message.get_payload
methods to see my cheat.

I also fixed a problem with the packet engine.  It wasn’t applying the
user mask on root ward messages.  The result was the cell agent getting
messages it was sending root ward.