More changes related to shift from error_chain to Failure
I had to handle errors between CellAgent and Message specially because
they formed a cycle.  That’s not the case for Failure, so I went back
to ? instead of an explicit match.

I also realized that the packetizer never returns an Error, so I got
rid of the Result return type.