A failed attempt to deploy multiple services on a single tree
I want to have a NocMaster service listening on a NocMaster tree and
NocAgents sending on that tree.  This version gets close to doing that,
but not quite.  I think the problem is the way it combines stacking a
tree and deploying services on it.  I think I need to split that up
into two steps but I’m not sure what pieces I can reuse.  Hence, I’m
saving a version that doesn’t work.