Added function to get &'static str from String
Which I need as a convenience when chaining errors with Failure.  The
object takes a &’static str, but sometimes I want to use a String.  I
could declare it String in the object,  but most of the time I’m
supplying an &’static str.  It’s a pain to remember to convert it to
String every time.