Getting "TCP" messages from Noc to cell agent
Previous versions had the NOC send EC messages to the port.  This
version changes things so a border port receives a non-EC message that
I’ve labeled TCP.

Each of these messages contains a message type and a
HashMap<String,String>, where the first string is a parameter name, and
the second string is a serialization of the parameter value.  The cell
agent uses these strings to construct an EC message on behalf of the
NOC.