Some changes to make stacking trees work
I fixed one of two problems with late DiscoverD messages.  Say that a
message gets forwarded by the packet engine on all the cell’s child
ports.  Then a message with Save=true is received, followed by a
DiscoverD adding a new child.  This version saves such a message to
send on the new child port when the traph is updated.

Unfortunately, there’s still a problem with a late DiscoverD.  Say that
a message with Save=true reaches the packet engine.  The packet engine
forwards the message to those children and to the cell agent.  If the
cell then processes a DiscoverD before the cell agent process the
message, that message never gets sent on the new child’s port.