StackTree works
I found the last race condition that was causing some cells not to see
the StackTreeMsg.  When we use packet forwarding, it is possible for a
message to be forwarded through a cell before a new child joins and
before the packet engine gets the new forwarding entry.  I now resend
any message with a save_eqn that evaluates to TRUE when the tree it was
sent on is updated.  That’s two places, update_traph and when the cell
agent receives the message from the packet engine.