Closed a possible security hole
The traphs data structure is a hash map indexed by the black tree ID,
which means I need that ID any time I need to access the traph, such as
when stacking a tree.  I had put the black tree ID in the stack tree
message, but that meant every service could name the black tree.

This version creates a map in the cell agent between the parent tree
ID, the tree the new tree is being stacked on, and the black tree at
the base of the stack.  Now the service does not need permission to
name the black tree in order to stack trees.