Trees stacked to deploy services
This version creates two trees for deploying the NocMaster and NocAgent
services and two trees for them to talk on, one for master->agent and
one for agent->master.  I had to fix an error in the way I evaluated
the gvm_equation to get the forwarding table entries right.