Change to Message Types, Manifest, and UUID
I don’t usually commit so many different changes at one time, but I
lost track of things.

It’s hard to debug when the names of things change on each run, as does
the UUID.  I created a fake UUID that returns the hash of the item ID.
That only works because these IDs are unique by construction.

I change the way I deploy containers.  Before I used the deploy step to
build the tree I was deploying on.  That’s awkward when I want the
container to have access to multiple trees.  I now stack the trees
first and then deploy on one of them.  In general, the container does
not get access to the tree it is deployed on.

I added a separate enum for message types to and from the outside.  The
main advantage is that I can use the compiler to catch missed cases.
The downside is that I have to explicitly return an error when the
outside sends a message only the inside should send.  Perhaps a later
version of the code will have separate enums for each direction.  I
didn’t do that for this version, because there is only one such message
type.