Tree Stacking closer
but still not working.  StackTreeMsg is reaching all the cells it
should, and I’m updating the other cell’s index for the parent.  I also
added a StackTreeMsgD so the parent can update the index entries for
its children on the new tree, but there is still a bug.

I’m committing this version because debugging will be easier if I fix
something in the error handling.  Right now I exit the listen loop on
an error, which generates a lot of error messages I don’t care about.
Instead, I’ll print the error message and go back to listening.