Changed handling errors in listen loops
Previously, when an error occurred in a listen loop running in a
thread, I’d return an error.  The result was a lot of “Listening on a
closed channel” errors, which made it hard to find the error I was
looking for.

This version recursively calls the function it’s in following an error
before the channel handles go out of scope.  The result is no closed
channel errors even if the thread dies.  Now it acts more like a
server, restarting the thread automatically.