VMs and Containers strarting
I can now stack deployment trees and deploy “containers” on them.  A
container can get messages through its VM to the cell agent, but the
cell agent isn’t forwarding them yet.

There is still work to do.  My method for tracking allowed trees is
based on UptreeID, but I only get the message from the VM, so I need to
change to VmID.

Also, neighbors of the border cell get the stack NocMaster tree request
on some runs.  Fortunately, this version does the right thing but for
the wrong reason.  I’m leaving it for now, but it needs to be fixed.