Previous checkin got messed up
I don’t know if I made untested changes before the checkin or if I
somehow got an older version of collagent.rs.  Whatever the case, it
doesn’t compile.

This version compiles and starts the VMs and containers.  It still
doesn’t deal with late child additions to the parent tree, but the
Master and Agents start on at least some of the cells.