Fixed handling of allowed trees
Each agent outside the cell agent (either a VM or something talking on
a border port) has a set of trees it is allowed to name.  I had tried
to use existing data structures as the index to get that set for a
given agent to no avail.  I should have remembered that I always get in
trouble when I overload something.

I finally introduced a SenderID to use as the index.  I can now verify
that the AllowedTree the agent specifies is one it’s allowed to use.