Container to Container messaging works
I can now stack a couple of trees, deploy a service on each tree, and
have one service send a message that reaches instances of the other
service.

Unfortunately, that doesn’t happen every time, because I haven’t dealt
with the case of the message reaching the packet engine before the cell
agent has sent the table entry for the tree the message was sent on.