Working 1/3 of the time
I can get container to container messages through, but 2/3 of the time
at least one cell runs into a problem.

The issue arises when a message comes in on a tree that the cell agent
knows about but the packet engine doesn’t yet.  In that case, the
message comes in on the cell’s control tree, but I need to look up the
traph information based on the tree the message was sent on.  The
easiest way to do that is to use the tree UUID in the packet header,
but I’m still trying to get rid of that field to save overhead.

The other approach is to add the tree ID to the message header.  That
means I have to update the message to send it on a different tree, but
I think that’s worth the tradeoff.  I’m committing this version in case
this approach is too cumbersome.