Getting messages container to container
Most of the time.  The border cell gets a request to stack a tree, then
one to deploy an up tree on that tree, and finally one to send an
application message.  The receiving containers print the message to the
console.

There are still two problems, one minor and one not.  The minor problem
is that some cells get the application message more than once.  That’s
almost certainly due to the way I save messages for late tree updates.

The major problem is that every once in a while, say 1 in 10 runs, a
cell gets the application message before it gets the message to deploy
the application.

I’m checking in this version because I made a lot of changes to get
this far.