Fixed a race condition
Part of the debugging problem I’ve been having is thinking I had one
message order race when I had two.  I just fixed one of them.  The
problem was updating the forwarding table for stacked trees when
getting a DiscoverD.  In fact, a stacked tree entry should only get
updated when receiving a StackTreeD or when recovering from a failover.
 Since I don’t have failover code in this version, I don’t know where
to put that update.  I’ve left a TODO in the code to remind me.