Fixed data structure updates
The bulk of my earlier race conditions came from not updating all the
data structures when a child was added to a stacked tree.  This version
fixes them.

There is still a race when a child is added to a stacked tree caused by
the update to the forwarding table and forwarding saved messages not
being atomic.  A sequence that causes trouble is

1. Receive a message on a tree
2. Receive notification of a new child on that tree.
3. Send a forwarding table update to the packet engine.
4. Send a saved message to the new child

The problem arises when a message arrives at the packet engine between
steps 3 and 4.  It currently gets forwarded, so it arrives ahead of a
saved message that was sent earlier.