Version with a message race
Even though I fixed a bug causing a message race for the previous check
in, there is still (at least) one left.  This one isn’t caused by a
bug.  It’s a result of a missing coordinate between the cell agent and
packet engine when a tree is being updated while messages on that tree
are in flight.  The document at ~/Dropbox (Earth Computing)/Earth
Computing Team Folder/Team/Alan/InOrder.pages explains the race and
several proposals to address the problem.

I’m checking in this version so I have a known state to return to as I
try to add the needed coordination.