Fixed double dispatch for StackTreeMsg and StackTreeDMsg
When I wrote the code in Message, I didn’t know that Rust did type
erasure on traits.  Hence, I split the work into those pieces done in
Message and those done in CellAgent.  The result was a lot of
complication trying to get the right payload out of the message from
within the cell agent, which only has the Message trait type.

This version fixes that problem for two of the message types by calling
a different method in the cell agent for each message type, passing the
message payload with the explicit type rather than the trait.