Fixed race condition
I was getting out of order messages when updating a tree while messages
on it were in flight.  The problem was with messages held by the cell
agent for delivery to a late arriving child.  Updating the routing
table entry before the cell agent could forward the messages let
messages sent later by the root race ahead.

The ideal solution would be to update the routing table entry only
after all messages held by the cell agent have been sent.  However, we
never know when the last one has been sent, because the packet engine
can send another one at any time.

The ideal solution would be to block messages on the offending port
only for the specific tree being update until the cell agent says to go
ahead.  The problem is that the port is a single queue for all trees.

One solution (Thanks, Atsushi.) is for the packet engine to keep a per
tree buffer and only block the port when the buffer gets full.

This version of the code does not implement this idea.  Instead, I
implemented a quick hack that has the packet engine block all ports
when receiving a message marked with a special flag bit set.  I’ve left
a TODO in the code to fix this problem later.