Some minor changes
If a tree skips some nodes, I can’t just set messages saved for late
child arrival on the tree with a mask, because the skipped nodes will
forward them to all their children resulting in duplicate receives.

Instead, I need to create a new table entry for those messages.  To
enable using that entry, this version adds a send_packets_by_index
method that gets the table index from the StackTreeD message instead of
the traph for the new tree.