Fixed problem with root ward messages
I wasn’t properly updating the index corresponding to the parent port
for stacked trees.  The fix means that StackTree messages do not get
forwarded by the packet engine.  They all go to the cell agent that
updates the other_index field in the message, which is similar to what
I do with Discover messages.