Fixed race condition created by previous fix
The fix for root ward messages meant that StackTree messages had to go
through the cell agent before being forwarded.  The problem is that I
wasn’t blocking the port until the message was forwarded.  The result
was manifest and application messages running ahead of the stack tree
messages.  This fix makes StackTreeMsg blocking, just as StackTreeDMsg
is.