Using a new feature in Rust 1.26
Before Rust 1.26 you could only use the “?” form of try in a method
that returns Result.  Since main.rs did not have a return value, main
had to call a function that did the real work.  The newest version of
Rust adds a return type to main.rs, so the extra function is no longer
needed.