Half way through trees with holes
We can build trees that have cells that can’t read messages passing
through them.  There is no need for them to send a StackTreeDMsg unless
one of their descendants can read messages on that tree.

The problem is with messages saved for late child arrival.  We need to
use a special routing table entry to make sure the forwarded messages
only go out on the port to the newly added cell.

This version doesn’t send a StackTreeDMsg if the cell can’t receive
messages on that tree, but sending those messages a descendent can
isn’t working.  That’s not a problem for the current use case, because
there are no such cells.