Create improved trace record format
I created a new format for trace records based on feedback from Bill
and John.  This one includes a type to make it easier to parse.

Unfortunately, there is no check for duplicate types, so an error won’t
be detected until the trace is analyzed.  I can check at runtime for
the model, but that approach won’t work in the distributed
implementation.