Added trace schema version to trace
It’s useful for a trace analyzer to know what version of the schema a
given trace is in.  This version adds a new first line to the trace
file.

We’ll probably have to come up with a better scheme in the distributed
case where different cells might be streaming trace records in
different versions.