Cmodel as pass through working
The first step of my refactoring replaced the cell agent and packet
engine talking directly to each other with the cmodel mediating the
communications.  In this version, the cmodel simply passes messages
along unchanged.  The next task will be to move packet operations from
the cell agent to the cmodel.