Save messages instead of packets
There are messages that need to be saved for late port connects
(Discover) and late child arrival (StackTree, Manifest, Application).
I was saving them as packets, since the cell agent was doing the
packetization.

That sounded like a premature optimization, but it turns out to have
been a wise decision based on what the compiler did to me when I
switched to saving messages.  For some reason, thread::spawn doesn’t
like Box<Message> even though the documentation says it should work.

This version uses a hack to get around the problem while still saving
and forwarding messages instead of packets.  Now I can move the
packetization into the cmodel.A