Finished some code cleanup
Took all references to packet engine out of cell agent and vice versa.
Removed references to message channels between them.

Also, the latest Rust update changed the behavior of to_owned().  It
used to be just like clone() in that it returned a value from a
reference.  Now it returns a reference, so I changed all to_owned() to
clone().

Along the way, I added the string “Tree:” to tree names, because it is
too easy to confuse them with cell names without it.