Packet forwarding works with hash table instead of index
I still have the table index in the code, but I believe I’m only using
hash table lookups when forwarding packets.  The true test will be when
I remove table index from the code.

There was one bug that was hard to find.  Some leaf cells on stacked
trees had a mask bit set for their parent port, which made the port
both a parent and a child.

The error was in process_stack_tree_msg in cellagent.rs where I was
setting up to forward saved messages.  Unfortunately, I ended up
setting the mask for the forwarding tree incorrectly.  That wasn’t a
problem when using the index, because I created a different one that I
hadn’t used yet.  However, I kept the tree_id for the new entry, so I
ended up with the wrong mask in a table entry I was actually using.