Added Baran Decentralized edge list
Bill provided an edge list for the Baran decentralized graph, which is
now included as an option.  Unfortunately, the original had the cells
in some edges in the wrong order, i.e., larger cell number first.  I
added code to allow for this.

I fixed a bug in traph.rs in the version without TableIndex.  I had
deleted one line too many, so I wasn’t updating the child list
correctly.

Now that I’m only using a hash table instead of an index, the entries
print out in an unpredictable order.  I added code to routing_table.rs
to print entries in the order they are created.