Use real UUIDs
Instead of fake UUIDs that are the same on every run, this version uses
real UUIDs that change.  It’s harder for debugging, but it exposed a
bug in my code that used fake UUIDs.

The fake UUIDs are the hash of the name field in the ID.  It turns out
that I was creating SenderIDs twice.  That didn’t matter with fake
UUIDs, because it was the same both times.  With real UUIDs, I got a
different value each time.  The fix was to remove the redundant
SenderID creation.

Another bug was related to a change in the name field of a TreeID.
Originally, the name field of a black tree was the same as that of the
root cell, which was confusing when reading trace records.  I added the
prefix “Tree:” to the name, which exposed another bug.  I was using the
name of a tree_id when constructing an AllowedTree, but that meant the
name field of the AllowedTree no longer matched the string provided in
the manifest.  The change was to pass the AllowedTree provided by the
sender of the StackTreeMsg.