Use implicit direction
Before this update the packet engine used information contained in a
packet to know the direction the packet was going.  Now the direction
is implicit based on the incoming port.  If that port is the parent,
the packet is sent leaf ward.  If the port is not the parent, the
packet is sent root ward, which means cells on the tree between the
root and the sender see the messages. (I use the term “to root” to
denote the case where only the root receives the message.) This version
has only the root see the message by having a separate tree on which
only the root may receive messages.  That’s a GVM equation and a
stacked tree that wouldn’t be needed if we had a “to root” option.