Removed direction info from packet and renamed uuid_fake
The new design determines packet direction implicitly.  If it came in
on a parent, it goes leaf ward.  If it came in on a child, it goes root
ward.  That means a given packet can go either direction, so it makes
no sense to encode direction in the packet.

I defined a uuid_fake.rs so I could have a UUID that was the same on
every run as an aid in debugging.  The plan always was to go back to
UUID.  However, we may want to encode some things, such as direction,
in the UUID.  I decided to keep my own module that right now just
returns the result from the UUID crate.  This way it will be easier to
make changes, but the name uuid_fake is misleading.  This version
renames it to uuid_ec.rs.