Encode packet type in CFTI
The Cellular Fabric Tree Identifier (CFTI) is 128 bits, the same size
as a UUID.  However, we have several packet types.  We could encode
them in the packet payload, but I don’t want the packet engine looking
there.

Instead, I stole a byte from the UUID and defined a set of codes.  If
it’s a normal packet, the packet simply gets forwarded based on the
CFTI in the forwarding table.  If it’s an AIT token, the packet engine
decides what to do based on the type, AIT, TECK, TACK, TOCK, or TICK.
In addition, there’s a time direction, FORWARD or REVERSE, for
successful and failed AIT transfer, respectively.