Ignore confinement for (MVD) minimum viable demo
Allowing applications to create trees rooted on the requesting cell is
a bit tricky if we try to confine them.  The problem is that only the
black tree can use links pruned in the trees the application is
permitted to use.  That’s quite limiting, but we can’t let applications
use the black tree without allowing them to increase the set of cells
they can talk to.

Rather than implementing a mechanism to deal with that, the MVD gives
every application permission to send messages on the black tree rooted
on itself.  I’ve bracketed the offending line with  TODOs because it’s
a big security hole if we don’t remove it before shipping.