Propagate AIT status through container and VM
While the cell agent knew if an application message was AIT or not, the
application, e.g., Service, had no way of communicating that it wanted
to send a message via AIT nor knowing that a received message was sent
that way.

I also fixed the handling of UUIDs.  I had just the tree UUID encoded,
but that complicated printing others.  Now they are all NORMAL unless
it’s an AIT TreeID.