Preparing to add failover
This version sets up some of the mechanism needed to add failover to
root ward.  The key idea is to track the root port number on each
Discover/DiscoverD to allow one branch to failover to a node on another
branch and still return messages to the original path.  The actual
failover isn’t implemented yet.