Changed quenching algorithm
In order to properly use port tree IDs, I need to make sure that each
cell sees each DiscoverMsg from each of the root ports.  The result is
a discover process that scales as V^2d^2 as opposed to the V^2d scaling
of simple quenching.  (Here V is the number of nodes and d is the
average degree of the nodes.)

The result is a noticeably slower Discover process and much bigger
forwarding tables.