Simple failover sets new parent forwarding table entry
The new parent in a successful simple failover now gets a forwarding
table for the failed branch.  I changed traph::new_element() to
traph::update_element() and introduced a new traph::new_element() that
is currently used only for the failover case.  There’s a lot of
repeated code that I need to get rid of.

I also made changes to how I create a PortNumber from a PortNo.  Plus,
I had to update utility.rs to handle a change in the uuid crate
interface.