Some minor fixes
I fixed a bug in collecting the trees with broken parents.  I was doing
a find while building the Vec, but that could miss some.  I’ve now
added an additional map to collect the values before I filter.

I changed the way I find the port tree in update_traph.  I had been
selecting the first element of the port_tree Vec, but that means I have
to understand the data structure.  Instead, I made the port tree part
of the traph data structure and return it every time I add a new port
tree.

In cellagent.rs, I replaced a match statement with an or_insert_with,
which is more compact and Rustier.