Fixed a kludge
I originally didn’t have path information in DiscoverDMsg, so I just
set the path for a child to 0.  I forgot to change that back when I
added path to the message.  That’s now fixed.

I also fixed a dangling debug print in packet_engine.rs.  The print
statement was under the wrong “if” condition.